Tried to handle floating-point rounding... gave up

+++ printing.h
@@ -39,7 +39,7 @@ void	print_char(t_array *output, t_arg arg, unsigned char uc);
 void	print_wchar(t_array *output, t_arg arg, wchar_t wc, int *error);
 void	print_string(t_array *output, t_arg arg, char *string);
 void	print_wstring(t_array *output, t_arg arg, wchar_t *string, int *error);
-void	print_float(t_array *output, t_arg arg, double value);
+void	print_float(t_array *output, t_arg arg, long double value);
 void	print_invalid(t_array *output, t_arg arg, va_list list, int *error);
 void	print_percent(t_array *output, t_arg arg, va_list list, int *error);
